refactor: move sidebar id contants into shared folder
