+++ src/app/app.js
@@ -5,23 +5,24 @@ import { ContextSelection } from '../context-selection/index.js'
 import {
     DataWorkspace,
     DataDetailsSidebar,
-    contextualHelpSidebarId,
-    dataDetailsSidebarId,
-    validationResultsSidebarId,
-    ValidationResultsSidebar,
 } from '../data-workspace/index.js'
 import {
     RightHandPanel,
     useRightHandPanelContext,
 } from '../right-hand-panel/index.js'
+import {
+    contextualHelpSidebarId,
+    dataDetailsSidebarId,
+    validationResultsSidebarId,
+} from '../shared/index.js'
 import { Layout } from './layout/index.js'
 import LoadApp from './load-app.js'
 // import { MutationIndicator } from './mutation-indicator/index.js'
 
 const idSidebarMap = {
+    [contextualHelpSidebarId]: ContextualHelpSidebar,
     [dataDetailsSidebarId]: DataDetailsSidebar,
     [validationResultsSidebarId]: ValidationResultsSidebar,
-    [contextualHelpSidebarId]: ContextualHelpSidebar,
 }
 
 /**
