+++ src/bottom-bar/data-item-bar.js
@@ -1,9 +1,8 @@
 import i18n from '@dhis2/d2-i18n'
 import { Button } from '@dhis2/ui'
 import React from 'react'
-import { dataDetailsSidebarId } from '../data-workspace/constants.js'
 import { useRightHandPanelContext } from '../right-hand-panel/index.js'
-import { useHighlightedField } from '../shared/index.js'
+import { dataDetailsSidebarId, useHighlightedField } from '../shared/index.js'
 import styles from './data-item-bar.module.css'
 
 export default function DataItemBar() {
