+++ src/bottom-bar/index.js
@@ -1,1 +1,2 @@
 export { default as BottomBar } from './bottom-bar.js'
+export * from './validation-results-sidebar/index.js'
