+++ src/bottom-bar/main-tool-bar.js
@@ -4,12 +4,12 @@ import { useIsMutating } from '@tanstack/react-query'
 import cx from 'classnames'
 import React from 'react'
 import { MutationIndicator } from '../app/mutation-indicator/index.js'
-import { validationResultsSidebarId } from '../data-workspace/constants.js'
 import useRightHandPanelContext from '../right-hand-panel/use-right-hand-panel-context.js'
 import {
     useDataValueSetQueryKey,
     useLockedContext,
     useEntryFormStore,
+    validationResultsSidebarId,
 } from '../shared/index.js'
 import styles from './main-tool-bar.module.css'
