+++ src/data-workspace/constants.js
@@ -3,7 +3,3 @@ export const FORM_TYPES = {
     SECTION: 'SECTION',
     CUSTOM: 'CUSTOM',
 }
-
-export const dataDetailsSidebarId = 'data-details'
-export const validationResultsSidebarId = 'validation-results'
-export const contextualHelpSidebarId = 'contextual-help'
