--- src/right-hand-panel/constants.js
@@ -1,1 +1,0 @@
-export const portalElementId = 'right-hand-panel-portal-anchor'
