+++ src/shared/constants.js
@@ -1,0 +1,4 @@
+export const completionSidebarId = 'completion'
+export const dataDetailsSidebarId = 'data-details'
+export const validationResultsSidebarId = 'validation-results'
+export const contextualHelpSidebarId = 'contextual-help'
