+++ src/shared/index.js
@@ -1,4 +1,5 @@
 export * from './data-value-mutations/index.js'
+export * from './constants.js'
 export * from './error-message/index.js'
 export * from './fixed-periods/index.js'
 export * from './highlighted-field/index.js'
